Move rehype-wrapify to src/
